fix: add sealed check before editing SDK config (#38)
<!-- < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < ☺
v                               ✰  Thanks for creating a PR! ✰    
v    Before smashing the submit button please review the checkboxes.
v    If a checkbox is n/a - please still include it but + a little note why
☺ > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > >  -->

## Description
Inside the Cosmos SDK there is a singleton object defined as `sdk.Config` that contains all the configuration for a chain. This object is also used inside Juno because it is used to store the Bech32 prefixes that are later used to serialize and deserialize addresses. Unfortunately, once that object is sealed we can no longer perform any operations on it. Since there are some chains that might configure it and seal it using an `init` method (such as Regen), we need to make that it is not sealed before setting up the addresses prefix. However, the variable that tells whether the config is sealed or not is private, which makes it impossible to know easily. For this reason, we implemented the usage of [reflection](https://en.wikipedia.org/wiki/Reflective_programming) in order to access the private `sealed` method and avoid setting it up if not possible. 

## Checklist
- [x] Targeted PR against correct branch.
- [ ] Linked to Github issue with discussion and accepted design OR link to spec that describes this work.
- [ ] Wrote unit tests.  
- [ ] Re-reviewed `Files changed` in the Github PR explorer.