feat: allow to handle MsgExec instances properly (#75)
## Description
This PR adds the possibility of handling `MsgExec` inner messages. 

Currently when a `MsgExec` is executed, the inner messages are not parsed correctly by the various modules. With this PR a new `AuthzMessageModule` interface is added. Thanks to this, modules will be able to properly handle `MsgExec` inner messages and parse their data accordingly.

## Checklist
- [x] Targeted PR against correct branch.
- [ ] Linked to Github issue with discussion and accepted design OR link to spec that describes this work.
- [ ] Wrote unit tests.  
- [x] Re-reviewed `Files changed` in the Github PR explorer.